\name{photoSpec-package}
\alias{photoSpec-package}
\alias{photoSpec}
\docType{package}
\title{
photoSpec: A photographic spectrophotometer
}
\description{
photoSpec allows one to use a photograph as a spectrophotometer, much in the manner that one might match the color of an object to a paint color at a hardware store.  Functions are provided to plot the CIE chromaticity diagram, select calibration colors, print a sample card containing calibration colors and a space for a sample to be photographed, and to analyze the results.  The motivating idea is to apply the concepts of Beer's Law to objects that cannot be chemically analyzed.
}
\details{
\tabular{ll}{
Package: \tab photoSpec\cr
Type: \tab Package\cr
Version: \tab 0.4-0\cr
Date: \tab 2013-06-15\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{ package }
