\name{runAllCal}
\alias{runAllCal}
\title{
Run all calibration options
}
\description{
This function runs both \code{\link{showCIEcalibration}} and \code{\link{showRGBcalibration}} and it extracts just the R, G and B channels if \code{sampCols} is provided.  The results are returned in a single data frame.
}
\usage{
runAllCal(calCols, sampCols = NULL, title = NULL, opts = "sRGB", ...)
}
\arguments{
  \item{calCols}{
An list with elements \code{rgb} and \code{hexcol}.  See \code{\link{genCalCols}} for details.
}
  \item{sampCols}{
A data frame with columns \code{hex} which contains hexadecimal strings, and \code{id} which has the sample names.
}
  \item{title}{
A character vector giving a title for the CIE plot.
}
  \item{opts}{
Character vector passed to \code{\link{plotCIEchrom}}.
}
  \item{\dots}{
Other options to be passed to either \code{grid} plotting functions, for \code{\link{showCIEcalibration}}, or \code{rgl} plotting functions, for \code{\link{showRGBcalibration}}.  Arguments must be appropriate for the plotting system.  See those functions for the possible options, along with the examples below.
}
}
\value{
Two plots are drawn, and information about the calibration is returned in a data frame if \code{sampCols} is provided.  The data frame is \code{sampCols} to which has been appended the purity and apparent lambda max (appLmax), as generated by \code{\link{showCIEcalibration}}.  \code{\link{showRGBcalibration}} appends the value and residual.  In addition, the values for the red, green and blue channels of \code{sampCols$hex} are appended.  Finally, if \code{lambdas} are provided, they are passed to \code{\link{showCIEcalibration}} and the results are appended to the data frame as well.
}
\references{
\url{https://github.com/bryanhanson/photoSpec}
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\examples{
# Some data
sc <- data.frame(hex = c("#FF0000", "#0000FF", "#98F5FF", "#6495ED"),
	id = c("red", "blue", "cadetblue1", "cornflower"))
blues <- genCalCols(minHue = "2.5B", maxHue = "10B", plotPC = FALSE)
#
tst1 <- runAllCal(calCols = blues, sampCols = sc)
tst2 <- runAllCal(calCols = blues, sampCols = sc, lambdas = 450, plotPts = TRUE, plotLambdas = TRUE)
}
\keyword{ plot }
