\name{photoSpec-package}
\alias{photoSpec-package}
\alias{photoSpec}
\docType{package}
\title{
photoSpec: A photographic spectrophotometer
}
\description{
photoSpec allows one to apply the concept of Beer's Law to a photograph containing color swatches (paint chips) and a sample to be analyzed.  The motivation is to provide a method for 'spectrophotometry' of objects that cannot be chemically analyzed because of their value or non-extractability.  Functions in the package cover the entire workflow from generating calibration colors to obtaining a quantitative answer analogous to concentration in Beer's Law.
}
\details{
\tabular{ll}{
Package: \tab photoSpec\cr
Type: \tab Package\cr
Version: \tab 0.5-2\cr
Date: \tab 2014-02-08\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\references{
\url{https://github.com/bryanhanson/photoSpec}
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{ package }
