\name{photoSpec-package}
\alias{photoSpec-package}
\alias{photoSpec}
\docType{package}
\title{
photoSpec: A photographic spectrophotometer
}
\description{
PENDING
}
\details{
\tabular{ll}{
Package: \tab photoSpec\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2013-04-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{ package }
