\name{prepCIEgradient}
\alias{prepCIEgradient}
\title{
Compute a gradient for display in a CIE chromaticity diagram
}
\description{
This function computes a gradient for display in a CIE chromaticity diagram.
}
\usage{
prepCIEgradient(vertices, colSpace, ff)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vertices}{
A data frame containing the xy coordinates of the vertices.
}
  \item{colSpace}{
The color space to use.
}
  \item{ff}{
The fudge factor to use
}
}
\details{
DOCUMENTATION NEEDS WORK!
}
\value{
A raster object as an x,y matrix of colors.
}
\references{
NEED A REFERENCE
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}

\keyword{ utilities }
