\name{selectCIExy}
\alias{selectCIExy}
\title{
Select a region of the CIE chromaticity diagram.
}
\description{
NEED A DESCRIPTION
}
\usage{
selectCIExy(L1, L2)
}
\arguments{
  \item{L1}{
Numeric, giving the wavelength at one edge of the region to be analyzed.
}
  \item{L2}{
Numeric, giving the wavelength at one edge of the region to be analyzed.
}
}
\details{
MAY NEED DETAILS
}
\value{
A vector of colors that fall in the polygon specified by the wavelength range.  Currently, this is sampled down to just 10 values.
}
\references{
NEED A REFERENCE
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\keyword{ plot }
