\name{selectCIExy}
\alias{selectCIExy}
\title{
Select a region of the CIE chromaticity diagram.
}
\description{
NEED A DESCRIPTION
}
\usage{
selectCIExy(lambda1, lambda2)
}
\arguments{
  \item{lambda1}{
Numeric, giving the wavelength at one edge of the region to be analyzed.
}
  \item{lambda2}{
Numeric, giving the wavelength at one edge of the region to be analyzed.
}
}
\details{
MAY NEED DETAILS
}
\value{
A data frame containing the vertices of a polygon which contains...
}
\references{
NEED A REFERENCE
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ plot }
